Allow decimal precision for oxalate_mg [#101343394]

+++ db/migrate/20150815220227_add_oxalate_mg_and_serving_to_foods.rb
@@ -1,6 +1,6 @@
 class AddOxalateMgAndServingToFoods < ActiveRecord::Migration
   def change
-    add_column :foods, :oxalate_mg, :integer, null: false, default: 0
+    add_column :foods, :oxalate_mg, :float, null: false, default: 0.0
     add_column :foods, :serving, :string
   end
 end
